git-pyfunction/requirements.txt
  # An example requirements file. If your function has other dependencies,
  # add them below
  functions-framework==3.*
  meraki
  firebase_admin
  python-dotenv
  requests
  httpx
